Refactor Calculator component & move Screen & Button in its own file
